+++ packages/manager/modules/pci/src/components/project/compute/infrastructure/vrack/index.js
@@ -3,8 +3,6 @@ import '@ovh-ux/ng-ovh-cloud-universe-components';
 import '@ovh-ux/ng-ovh-api-wrappers';
 import 'ovh-api-services';
 
-import projectPrices from '../../../prices';
-
 import factory from './factory';
 import virtualMachineFactory from './virtualMachine/factory';
 import vlanFactory from './vlan/factory';
@@ -14,7 +12,6 @@ const moduleName = 'ovhManagerPciComponentsProjectComputeInfrastructureVrack';
 angular
   .module(moduleName, [
     'ovh-api-services',
-    projectPrices,
     'ngOvhCloudUniverseComponents',
   ])
   .factory('CloudProjectComputeInfraVrackFactory', factory)