+++ packages/manager/modules/config/CHANGELOG.md
@@ -1,3 +1,12 @@
+# [0.1.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-config@0.1.0-alpha.0...@ovh-ux/manager-config@0.1.0) (2019-03-19)
+
+
+### Bug Fixes
+
+* remove prerelease state ([8708ce0](https://github.com/ovh-ux/manager/commit/8708ce0))
+
+
+
 # [0.1.0-alpha.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-config@0.0.0...@ovh-ux/manager-config@0.1.0-alpha.0) (2019-03-19)