+++ packages/manager/modules/sms/CHANGELOG.md
@@ -1,3 +1,26 @@
+# [6.0.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-sms@5.0.0...@ovh-ux/manager-sms@6.0.0) (2019-03-19)
+
+
+### Code Refactoring
+
+* bump all packages to [@ovh-ux](https://github.com/ovh-ux)/manager-core@^5.0.0 ([7cbc70a](https://github.com/ovh-ux/manager/commit/7cbc70a))
+
+
+### BREAKING CHANGES
+
+* Until theses packages has a dependency to @ovh-ux/manager-core@^5.0.0, the host project needs to import @ovh-ux/manager-config
+
+Before:
+
+yarn add @ovh-ux/manager-core
+
+Now:
+
+yarn add @ovh-ux/manager-config
+yarn add @ovh-ux/manager-core
+
+
+
 # [5.0.0](https://github.com/ovh-ux/manager/compare/@ovh-ux/manager-sms@4.0.0...@ovh-ux/manager-sms@5.0.0) (2019-03-14)